requirements.txt
# Smart Traffic Light Monitoring System - Dependencies
# Compatible with Python 3.13

# Core Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep Learning - YOLOv8
ultralytics>=8.0.0
# PyTorch (already installed, but keeping for reference)
# torch>=2.0.0
# torchvision>=0.15.0

# Data Processing
numpy>=1.26.0  # Python 3.13 compatible
pandas>=2.1.0

# Image Processing
Pillow>=10.0.0

# Configuration
PyYAML>=6.0

# Logging and Monitoring
colorlog>=6.7.0

# GUI (optional)
matplotlib>=3.8.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Utilities
tqdm>=4.66.0
scipy>=1.11.0

# Testing
pytest>=7.4.0

# ===== NEW FEATURES =====

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Email Alerts
# (Already included in Python standard library: smtplib, email)

# Report Generation
fpdf2>=2.7.0  # PDF reports
openpyxl>=3.1.0  # Excel reports
xlsxwriter>=3.1.0  # Excel formatting
